externalize the NetStringServerSocket class from the Server class
